user32: Fix messages sent on a window without WS_CHILD, but with an invisible parent, when it is shown.
Some applications depend on the fact that WM_WINDOWPOSCHANGING is sent
after WM_SHOWWINDOW when SetParent is called, even if the window ends up
not visible because its parent is not visible, and occurs when the window
itself does not have the WS_CHILD style set.

This also fixes a TODO message sequence, so that Wine matches Windows.

Wine-Bug: https://bugs.winehq.org/show_bug.cgi?id=40262
Signed-off-by: Gabriel Ivăncescu <gabrielopcode@gmail.com>