fixup! user32: Use the desktop shared data for GetCursorPos().
Check the window station and thread desktop handle validity in get_desktop_shared_memory().
When starting a new thread, the thread desktop may be NULL and thus NtUserGetObjectInformation()
will fail and then try to open a memory section with uninitialized name buffer. This saves
a few server calls when unnecessary.